fix: make download.sh less destructive, change defaults in web, fix redirect
Added database checking that works, so any database file name can be
used, made sure .env is NOT overwritten if it exists.

Edited the api_base_url in petring-web.toml to be localhost rather than
the full on domain

Changed the redirect location from /get/users/{username}/* to
/get/user/{username} so that redirects dont lead to a 404

feat: graceful shutdown of both servers

feat: cleanup routes, run fmt, add new routes, improve bulk delete

Cleaned up and moved route declarations for both web and api to new
locations to not clog up main.

Configured rustfmt to use 2 spaces instead of 4, cause I just prefer it
that way, formatting is opinionated.

Added 2 new routes get_ad and get_ads_unverified to help in proper
moderation.

Improved bulk delete by batch querying instead of iterating

fix: remove stupid refresh block

fix: docker-compose files

feat: add .editorconfig

+++ download.sh
@@ -7,36 +7,36 @@ shopt -s nullglob
 set -o pipefail
 
 function handle_error() {
-    local exit_code="$1"
-    local message="$2"
+  local exit_code="$1"
+  local message="$2"
 
-    echo "Error: ${message}"
+  echo "Error: ${message}"
 
-    exit "$exit_code"
+  exit "$exit_code"
 }
 
-
+DB_FILE="./petring.db"
 NO_JUST=false
 
 if ! command -v git &> /dev/null; then
-    echo "git is not installed"
-    exit 1
+  echo "git is not installed"
+  exit 1
 fi
 
 if ! command -v cargo &> /dev/null; then
-    echo "cargo is not installed"
-    echo "cargo is required to run migrations"
-    exit 1
+  echo "cargo is not installed"
+  echo "cargo is required to run migrations"
+  exit 1
 fi
 
 if ! command -v sqlite3 &> /dev/null; then
-    echo "sqlite3 is not installed"
-    exit 1
+  echo "sqlite3 is not installed"
+  exit 1
 fi
 
 if ! command -v just &> /dev/null; then
-    echo "just is not installed, trying without it"
-    NO_JUST=true
+  echo "just is not installed, trying without it"
+  NO_JUST=true
 fi
 
 echo "Cloning repo"
@@ -44,30 +44,37 @@ git clone https://github.com/h4rldev/petring.git ./buf
 
 echo "Getting relevant files"
 mv -t ./ ./buf/migration ./buf/frontend ./buf/petring-api.toml ./buf/petring-web.toml ./buf/docker-compose.yml ./buf/docker-compose.override.yml
-mv ./buf/.env.example ./.env
+
+if [[ ! -f ./.env ]]; then
+	mv ./buf/.env.example ./.env
+fi
+
 if [[ ${NO_JUST} == false ]]; then
-    mv ./buf/justfile .
+  mv ./buf/justfile .
 fi
 
 echo "Removing cloned repo"
 rm -fr ./buf
 
 for file in ./*.db; do
-    if [[ ! -f "$file" ]]; then
-        echo "Generating empty database"
-        sqlite3 petring.db "VACUUM;"
-    fi
+  if [[ ! -f "$file" ]]; then
+    echo "Generating empty database"
+    sqlite3 petring.db "VACUUM;"
+  fi
+
+	file_name=$(basename "$file")
+	DB_FILE="./$file_name"
 done
 
 echo "Running migrations"
 if [[ ${NO_JUST} == true ]]; then
-    pushd ./migration/ >/dev/null || handle_error 1 "Failed to pushd to migrations"
-    mv ../petring.db .
-    cargo run --release -- up
-    mv petring.db ../
-    popd >/dev/null || handle_error 1 "Failed to popd from migrations"
+  pushd ./migration/ >/dev/null || handle_error 1 "Failed to pushd to migrations"
+  mv ../"$DB_FILE" ./
+  cargo run --release -- up
+  mv "$DB_FILE" ../
+  popd >/dev/null || handle_error 1 "Failed to popd from migrations"
 else
-    just migrations up
+  just migrations up
 fi
 
 echo "Petring is now ready to be configured"
